type cating

+++ examples/symmetric_conv2d_vae/updated/lightning.py
@@ -60,7 +60,7 @@ class CVAE(pl.LightningModule):
     def training_step(self, train_batch, batch_idx):
         # x, y = train_batch
         x = train_batch["X"]
-        return {'forward': self.forward(x), 'expected': x}
+        return {'forward': self.forward(x).half(), 'expected': x.half()}
 
     def training_step_end(self, outputs):
         _, recon_x = outputs['forward']
@@ -76,7 +76,7 @@ class CVAE(pl.LightningModule):
 
     def test_step(self, test_batch, batch_idx):
         x = train_batch["X"]
-        return {'forward': self.forward(x), 'expected': x}
+        return {'forward': self.forward(x).half(), 'expected': x.half()}
 
     def test_step_end(self, outputs):
         _, recon_x = outputs['forward']
